Tab completion in the REPL for MMCIFDict and MMTFDict (#27)
* Add support for tab-completion of MMCIFDict keys

This is achieved by subtyping from `AbstractDict`.  The methods
`Base.length` and `Base.iterate` were needed to make `Base.show` work
for MIME("text/plain") output.

* Add test for Base.show for MMCIFDict

* Add support for tab-completion dor MMTFDict keys

The functions `Base.length` and `Base.iterate` are needed
for `Base.show` to work.  This didn't turn up when running
the tests, so a simple test that `Base.show` works was
added.