Add van der Waals radii data from Mantina et al.
When rendering structures or calculating interactions, it's useful to
have van der Waals radii for each element. This commit adds a dictionary
of van der Waals radii sourced from PMID: 19382751 to the bonding.jl
file. This is most appropriate when the structure has added hydrogens.
(These are NOT the "united atom radii" of PMID: 10388571 that attempt to
account for missing hydrogens.)

The data are drawn from MolecularGraph.jl, added by me in
https://github.com/mojaie/MolecularGraph.jl/pull/47.